[server] Do not truncate cpu stats from double to float

+++ src/engine/server/sv_ccmds.cpp
@@ -243,7 +243,7 @@ public:
 			return;
 		}
 
-		float cpu = ( svs.stats.latched_active + svs.stats.latched_idle );
+		double cpu = ( svs.stats.latched_active + svs.stats.latched_idle );
 
 		if ( cpu )
 		{
